Update icon and switch to round add button and swipe actions
Change the icon from the generic Xamarin icon. Switch the add button on the main page
to be a round button "wow such fashion". Change edit and complete to be swipe actions.
Add an undo action as the inverse of the complete action in case you actually complete
the wrong thing.