Refactor the BaseDatabase and GoalDatabase classes
The way that I was ensuring the table creation on every call to get the
connection was resulting in ugly code. Now we'll ensure table creation
in the App.OnStart only. Eventually, I may need to consider how to version the
database, but that's a problem for the future. I also added in missing AttemptAndRetry()
and ConfigureAwait(false) calls so that all DB interactions are consistent.